refactor(l5): move logger from src/ into server/logging/legacy-logger behind a settings reader and an L5 adapter (issue 864)
L5 of _plans/server-hardening-ladder.md, issue 864: cut server/ free of src/.

Moved (git mv, then split at the seams the file already had):

  src/logger.ts -> server/logging/legacy-logger.ts

The 592-line original is now four siblings under server/logging/, each under
the whole-file lint standard:

  legacy-logger.ts           level state, context reader, extra sinks,
                             envelope, emit, sink-throw reporting
  legacy-logger-identity.ts  host name, service name, rotating file sink,
                             deriveWorkerLogPath
  legacy-logger-redaction.ts redactForLog and the marker-density guard
  legacy-logger-settings.ts  the settings reader contract

COLLISION NOTE: server/logging/logger.ts already existed and is a DIFFERENT
API (createLogger, withOperation, the pino-shaped one server/main.ts builds).
The two are not merged and do not import each other; the moved module is
named legacy-logger to keep that distinction visible.

ENVIRONMENT READS (8, all removed from the moved module — no process.env
under server/):

  line  key                       scope         now
   12   LOG_LEVEL                 module        reader, first use
  116   HOSTNAME                  module        reader, first use
  116   HOST                      module        reader, first use
  176   LOG_FILE                  module        reader, first use
  180   OPEN_BRAIN_WORKER_NAME    module        reader, first use
  184   LOG_MAX_BYTES             module        reader, first use
  185   LOG_MAX_FILES             module        reader, first use
  202   SERVICE_NAME              module        reader, first use
  203   OPEN_BRAIN_WORKER_NAME    module        reader, first use

Every read was at MODULE SCOPE in the original — the initial level, the host
label, the service label, and the file sink were all resolved once at import.
First-use is preserved by memoizing each on its first read instead of at
module load: getEffectiveLevel(), getHostName(), getServiceName() and
getFileSink() each resolve exactly once and cache, so the value is still read
once per process, just at the moment the first line is emitted rather than at
import. That deferral is what lets the adapter and server/main.ts register a
reader during import and still be the source of the value.

READER CONTRACT (rule M13a — a reader, not a snapshot, because the file suite
sets the environment of a spawned child after import):

  setLegacyLoggerSettingsReader(read: () => LegacyLoggerSettings)
    : () => LegacyLoggerSettings   // returns the previous reader

The default reader, installed before any registration, returns the
environment-absent defaults: level info, service open-brain, no file sink.
A script or test importing a server/ module directly therefore logs exactly
as it did with none of these variables set.

ADAPTER: src/logger.ts is now an M9 adapter (23 code lines). It parses the
environment with the original keys, precedence and defaults, registers that
parse as the reader at import, and re-exports everything. Every export keeps
its name and arity: setLogLevel, getLogLevel, deriveWorkerLogPath,
setLogContextReader, addLogSink, redactForLog, logger.

IMPORTERS UNCHANGED (rule M3): the 24 server/ importers of ../../src/logger.ts
and every src/ and scripts/ importer keep their current import line — they hold
no configuration, so the repoint belongs to L6.

server/main.ts: one import plus one call in startServer, immediately after
`const config = ...` and before the first line can be emitted, registering a
reader over the values it already holds (config.logging.level/.file/.service/
.workerName and config.transport.hostname). server/config.ts is untouched: it
already carries a logging group with those four fields. The rotate threshold
and retained-file count are not in config and stay undefined, which is the
value the module already defaults on.

.oxlintrc.json: server/logging/legacy-logger.ts joins the existing no-console
override beside src/logger.ts and server/logging/logger.ts. That override
follows the file rather than being a new exemption — this module IS the thing
that wraps the console sink, which is why the entry exists.

Two functions were hoisted to stay under the complexity rule, with the
branches, their order and their outputs unchanged: tameValue() out of the
JSON.stringify replacer, and describeSinkThrow() out of reportSinkThrow().
The sink snapshot spread became Array.from with identical semantics.

Tests: src/logger-file.test.ts stays at its src/ path untouched (rule M1
EXCEPTION) and exercises the moved code through the adapter — it spawns a
child with a constructed environment, which is exactly the case the reader
exists to keep working.

Closure count: 18 -> 17.
